[docs] all: Fix all lint and spelling issues
