+++ docs/gettingstarted/requirements.md
@@ -16,59 +16,30 @@ Moodle 4.0 supports the following PHP versions:
 
 The following PHP extensions are required (most of which are installed and enabled by default in most PHP installations):
 
-- [iConv],
-- [mbstring],
-- [curl],
-- [openssl],
-- [ctype],
-- [zip],
-- [zlib],
-- [simplexml],
-- [spl],
-- [pcre],
-- [dom],
-- [xml],
-- [xmlreader],
-- [intl],
-- [json],
-- [hash],
-- [fileinfo]
+- [`iconv`]: https://www.php.net/book.iconv
+- [`mbstring`]: https://www.php.net/book.mbstring
+- [`curl`]: https://www.php.net/book.curl
+- [`openssl`]: https://www.php.net/book.openssl
+- [`ctype`]: https://www.php.net/book.ctype
+- [`zip`]: https://www.php.net/book.zip
+- [`zlib`]: https://www.php.net/book.zlib
+- [`simplexml`]: https://www.php.net/book.simplexml
+- [`spl`]: https://www.php.net/book.spl
+- [`pcre`]: https://www.php.net/book.pcre
+- [`dom`]: https://www.php.net/book.dom
+- [`xml`]: https://www.php.net/book.xml
+- [`xmlreader`]: https://www.php.net/book.xmlreader
+- [`intl`]: https://www.php.net/book.intl
+- [`json`]: https://www.php.net/book.json
+- [`hash`]: https://www.php.net/book.hash
+- [`fileinfo`]: https://www.php.net/book.fileinfo
 
 ## Relational Database
 
 The following relational database servers are supported. The relevant PHP extension will also be required.
 
-- [MariaDB][mariadb] (version 10.2.29 or higher) with the [MySQLi PHP Extension][php.mysqli]
-- [MySQL][mysql] (version 5.7 or higher) with the [MySQLi PHP Extension][php.mysqli]
-- [Postgresql][postgres] (version 10 or higher) with the [pgsql PHP Extension][php.pgsql]
-- [Microsoft SQL Server][sqlsrv] (version 14.0 or higher) with the [SQLSRV PHP Extension][php.sqlsrv]
-- [Oracle][oci] (version 11.2 or higher) with the [OCI8 PHP Extension][php.oci8]
-
-[php]: https://php.net/downloads/
-[iconv]: https://www.php.net/book.iconv
-[mbstring]: https://www.php.net/book.mbstring
-[curl]: https://www.php.net/book.curl
-[openssl]: https://www.php.net/book.openssl
-[ctype]: https://www.php.net/book.ctype
-[zip]: https://www.php.net/book.zip
-[zlib]: https://www.php.net/book.zlib
-[simplexml]: https://www.php.net/book.simplexml
-[spl]: https://www.php.net/book.spl
-[pcre]: https://www.php.net/book.pcre
-[dom]: https://www.php.net/book.dom
-[xml]: https://www.php.net/book.xml
-[xmlreader]: https://www.php.net/book.xmlreader
-[intl]: https://www.php.net/book.intl
-[json]: https://www.php.net/book.json
-[hash]: https://www.php.net/book.hash
-[fileinfo]: https://www.php.net/book.fileinfo
-
-[mariadb]: https://mariadb.org/
-[php.mysqli]: https://www.php.net/manual/en/book.mysqli.php
-[mysql]: https://www.mysql.com/
-[postgres]: https://www.postgresql.org/
-[sqlsrv]: https://www.microsoft.com/en-au/sql-server/sql-server-downloads
-[oci]: https://docs.oracle.com/cd/B19306_01/server.102/b14220/intro.htm
-[php.oci8]: https://www.php.net/manual/en/book.oci8.php
-[php.pgsql]: https://www.php.net/manual/en/book.pgsql.php
-[php.sqlsrv]: https://www.php.net/manual/en/book.sqlsrv.php
+- [MariaDB](https://mariadb.org/) (version 10.2.29 or higher) with the [MySQLi PHP Extension](https://www.php.net/manual/en/book.mysqli.php)
+- [MySQL](https://www.mysql.com/) (version 5.7 or higher) with the [MySQLi PHP Extension](https://www.php.net/manual/en/book.mysqli.php)
+- [Postgresql](https://www.postgresql.org/) (version 10 or higher) with the [pgsql PHP Extension](https://www.php.net/manual/en/book.pgsql.php)
+- [Microsoft SQL Server](https://www.microsoft.com/en-au/sql-server/sql-server-downloads) (version 14.0 or higher) with the [`SQLSRV` PHP Extension](https://www.php.net/manual/en/book.sqlsrv.php)
+- [Oracle](https://docs.oracle.com/cd/B19306_01/server.102/b14220/intro.htm) (version 11.2 or higher) with the [OCI8 PHP Extension](https://www.php.net/manual/en/book.oci8.php)
