+++ docs/guides/javascript/_examples/promises/getModal.js
@@ -1,6 +1,6 @@
 const getModal = questionBody => {
     return ModalFactory.create({
-        title: getString('mytitle', 'mod_example'),
+        title: getString('my_title', 'mod_example'),
         body: renderTemplate('mod_example/example_body', questionBody),
         removeOnClose: true,
     })
