+++ project-words.txt
@@ -1,61 +1,61 @@
-mnet
+Atto
+Bangma
+Brickfield
+Dougiamas
+IDEs
+MDL
+MDLQA
+MDLSITE
 Moodle
+MoodleCloud
+Moodleisms
 Moodlelib
 Moodlers
-adminpresets
-codingstyle
-gradebook
-phpunit
-behat
-plugintype
-modname
-Dougiamas
-Poltawski
-MDL
-MDLSITE
-MDLQA
-opcache
-thirdparty
-Packagist
-hackfest
-Bangma
 Mudrak
-Moodleisms
-unhold
-MoodleCloud
-triagers
-contextmodule
-contextcourse
-contextcoursecat
-siteadmin
-frontpage
-capabilityname
-pluginname
-riskbitmask
-captype
-contextlevel
-editingteacher
-moodledemo
-frankenstyle
-XMLDB
+Mudrák
 MySQLi
-pgsql
+Netbeans
 OCI8
+PHPDoc
+PHPDocumentor
+Packagist
+Poltawski
+TODO
 Triaged
 Untriaged
-webservice
-webservices
+XMLDB
+adminpresets
+behat
+capabilityname
+captype
+codingstyle
+contextcourse
+contextcoursecat
+contextlevel
+contextmodule
+editingteacher
+frankenstyle
+frontpage
+gradebook
+hackfest
+mnet
+modname
+moodledemo
+moodledev
+opcache
+pgsql
 phpdoc
 phpstorm
 phptags
-Brickfield
-Atto
-Mudrák
-moodledev
+phpunit
+pluginname
+plugintype
+riskbitmask
 ruleset
-TODO
+siteadmin
+thirdparty
 todos
-Netbeans
-IDEs
-PHPDoc
-PHPDocumentor
+triagers
+unhold
+webservice
+webservices
